requirements.txt
schedule
openai
moviepy
requests
cloudinary
python-dotenv
streamlit
numpy
scipy